requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.9.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
Flask==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
img2pdf==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ocrmypdf==16.4.3
openai==1.35.14
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.2
pi_heif==0.18.0
pikepdf==9.1.2
pillow==10.4.0
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypdfium2==4.30.0
pytesseract==0.3.13
python-dotenv==1.0.1
rich==13.7.1
sniffio==1.3.1
tesseract==0.1.3
tqdm==4.66.4
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.16.0
